+++ example/lib/main.dart
@@ -146,7 +146,17 @@ class _DataPageState extends State<DataPage> {
         editable: true,
       ),
       DatatableHeader(
-          text: "Name", value: "name", show: true, flex: 2, sortable: true, editable: true, textAlign: TextAlign.left, format: DataTableFormat.number, textInputFormatter: []),
+        text: "Name",
+        value: "name",
+        show: true,
+        flex: 2,
+        sortable: true,
+        editable: true,
+        textAlign: TextAlign.left,
+        format: DataTableFormat.number,
+        withButtonClear: false,
+        textInputFormatter: [],
+      ),
       DatatableHeader(
         text: "SKU",
         value: "sku",